BhoomiAI-main/backend/requirements.txt
flask
flask-cors
numpy
pandas
scikit-learn
joblib

gunicorn
deep-translator
ollama-python

